Updated Temperature Functions.

+++ AVR_Datalogger/main.h
@@ -3,4 +3,4 @@ extern uint8_t channel;
 extern uint8_t gain;
 extern uint8_t uartMenu;
 extern uint8_t hasSDCard;
-extern uint16_t ambientTemperature;
+extern float ambientTemperature;
